requirements.txt
python-telegram-bot==21.6
aiohttp>=3.9
python-dotenv>=1.0
